refactor(preprod): create size directory for status checks
