+++ bin/preprod/trigger_size_status_check
@@ -12,7 +12,7 @@ logger = logging.getLogger(__name__)
 from sentry.models.commitcomparison import CommitComparison
 from sentry.models.project import Project
 from sentry.preprod.models import PreprodArtifact, PreprodArtifactSizeMetrics
-from sentry.preprod.vcs.status_checks.tasks import create_preprod_status_check_task
+from sentry.preprod.vcs.status_checks.size.tasks import create_preprod_status_check_task
 
 #################
 # Configuration #
