+++ tests/sentry/preprod/vcs/status_checks/size/test_status_checks_tasks.py
@@ -7,7 +7,7 @@ from sentry.integrations.source_code_management.status_check import StatusCheckS
 from sentry.models.commitcomparison import CommitComparison
 from sentry.models.repository import Repository
 from sentry.preprod.models import PreprodArtifact
-from sentry.preprod.vcs.status_checks.tasks import create_preprod_status_check_task
+from sentry.preprod.vcs.status_checks.size.tasks import create_preprod_status_check_task
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import region_silo_test
 
@@ -79,11 +79,11 @@ class CreatePreprodStatusCheckTaskTest(TestCase):
         mock_provider.create_status_check.return_value = "check_12345"
 
         patcher_client = patch(
-            "sentry.preprod.vcs.status_checks.tasks._get_status_check_client",
+            "sentry.preprod.vcs.status_checks.size.tasks._get_status_check_client",
             return_value=(mock_client, repository),
         )
         patcher_provider = patch(
-            "sentry.preprod.vcs.status_checks.tasks._get_status_check_provider",
+            "sentry.preprod.vcs.status_checks.size.tasks._get_status_check_provider",
             return_value=mock_provider,
         )
 
@@ -95,11 +95,11 @@ class CreatePreprodStatusCheckTaskTest(TestCase):
         mock_provider = Mock()
 
         patcher_client = patch(
-            "sentry.preprod.vcs.status_checks.tasks._get_status_check_client",
+            "sentry.preprod.vcs.status_checks.size.tasks._get_status_check_client",
             return_value=(mock_client, None),
         )
         patcher_provider = patch(
-            "sentry.preprod.vcs.status_checks.tasks._get_status_check_provider",
+            "sentry.preprod.vcs.status_checks.size.tasks._get_status_check_provider",
             return_value=mock_provider,
         )
